+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Version 7.0.0-dev
 - BC break: only array can be passed to attribute Skip (if any value is passed)
 - improved error message for assertions assertSame and assertNotSame when any passed value is boolean
 - deprecated method TAssertions::showStringOrArray in favor of new showValue (the latter accepts value of any type)
+- allowing skipping tests based on OS family with default skip checker
 
 Version 6.1.0
 - added CodeCoverageExtension, Reader, IAnnotationsReaderEngine, PhpAttributesEngine, TAssertions::getCounter(), TestsStartedEvent, TestsFinishedEvent to public api